v3

+++ app/main.py
@@ -2,16 +2,22 @@ import os
 
 
 def move_file(command: str) -> None:
-    conditions = command.split()
-    source_file = conditions[1]
-    final_file = conditions[2]
-    final_file_directories = final_file.split("/")[:-1]
+    command_name, source_file, final_file = command.split()
 
-    if len(final_file_directories) > 0:
-        os.makedirs("/".join(final_file_directories), exist_ok=True)
-    with (
-        open(source_file, "r") as start_file,
-        open(final_file, "w") as final_file
+    if (
+            command_name and
+            source_file and
+            final_file and
+            command_name == "mv"
     ):
-        final_file.write(start_file.read())
-    os.remove(source_file)
+        file_directories, file_name = os.path.split(final_file)
+
+        if file_directories:
+            os.makedirs(file_directories, exist_ok=True)
+
+        with (
+            open(source_file, "r") as start_file,
+            open(final_file, "w") as final_file
+        ):
+            final_file.write(start_file.read())
+        os.remove(source_file)
